Rename files, fix maps
